requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
django-filter==25.1
python-dotenv==1.0.0
requests==2.31.0
paystackapi==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
boto3==1.40.51
django-storages==1.14.6